New movement implementation?
Well it's an attempt anyway :P. Having some trouble with the movement
script.

Currently I'm able to make the player move at an even speed regardless
of distance from centre, but SOMETHING is failing causing the player to
movement to slowly spiral out of orbit. This has been giving me
headaches.

Also implemented is lane movement. It's currently a little sloppy, as
the values aren't exact at the moment. However, the movement itself has
a little bounce to it like in the mockup video, which is nice. The
script also tracks which lane the player in so that they aren't able to
move out of the rings.

Note: I'm sure some of this may read quite badly or fail to make sense.
I just realised 2am has passed and I'm incredibly drowsy. Night night i
guess.